feat: Split tree creation is a hook
