+++ src/components/Demonstration/index.jsx
@@ -1,6 +1,4 @@
 import React, { useCallback, useEffect, useMemo, useState } from 'react'
-import Input from 'cozy-ui/react/Input'
-import Label from 'cozy-ui/react/Label'
 import { useClient, useQueryAll } from 'cozy-client'
 import {
   nodesQuery,
@@ -10,9 +8,9 @@ import {
 import createTree from 'lib/createTreeExported'
 import TreeNetworkGraph from './TreeNetworkGraph'
 import Spinner from 'cozy-ui/react/Spinner'
-import { omit } from 'lodash'
 import Button from 'cozy-ui/transpiled/react/Buttons'
 import NodesTable from './NodesTable'
+import TreeCreator from './TreeCreator'
 
 const Demonstration = () => {
   const client = useClient()
@@ -27,16 +25,8 @@ const Demonstration = () => {
   )
   const supervisorWebhook = webhooks?.find(e => e.message.name === 'observe')
     ?.links.webhook
-  const [depth, setDepth] = useState(3)
-  const [fanout, setFanout] = useState(3)
-  const [groupSize, setGroupSize] = useState(2)
   const [lastExecutionId, setLastExecutionId] = useState()
   const [isRunning, setIsRunning] = useState(false)
-  const treeStructure = useMemo(() => ({ depth, fanout, groupSize }), [
-    depth,
-    fanout,
-    groupSize
-  ])
   const [tree, setTree] = useState([])
   const observationsQuery = observationsByExecutionQuery(
     tree ? tree[0]?.executionId : undefined
@@ -49,93 +39,18 @@ const Demonstration = () => {
     () => (rawObservations || []).filter(o => o.action !== 'receiveShare'),
     [rawObservations]
   )
-  const [treeNodes, treeEdges] = useMemo(() => {
-    if (!tree) return []
+  const [treeNodes, setTreeNodesState] = useState()
+  const [treeEdges, setTreeEdgesState] = useState()
+  const treeStructure = useMemo(() => {
+    return treeNodes && treeNodes[0] ? treeNodes[0].treeStructure : undefined
+  }, [treeNodes])
 
-    const usedProperties = node => {
-      return omit(node, ['parents'])
-    }
-
-    function onlyUnique(value, index, array) {
-      return array.indexOf(value) === index
-    }
-
-    const nodesMap = {}
-    const edgesMap = {}
-    // Converting the tree used for the execution (starting from its leaves) to D3 compatible data
-    const transformNode = node => {
-      // Initialize the current node
-      if (!nodesMap[node.nodeId]) {
-        nodesMap[node.nodeId] = { parents: [], children: [] }
-        edgesMap[node.nodeId] = []
-      }
-
-      // Converting parent(s)
-      for (const parent of node.parents || []) {
-        nodesMap[node.nodeId].parents = [
-          ...nodesMap[node.nodeId].parents,
-          parent.nodeId
-        ].filter(onlyUnique)
-        edgesMap[node.nodeId] = [
-          ...edgesMap[node.nodeId],
-          parent.nodeId
-        ].filter(onlyUnique)
-
-        transformNode(parent)
-        nodesMap[parent.nodeId] = {
-          ...nodesMap[parent.nodeId],
-          children: [...nodesMap[parent.nodeId].children, node.nodeId].filter(
-            onlyUnique
-          )
-        }
-      }
-      nodesMap[node.nodeId] = {
-        ...nodesMap[node.nodeId],
-        ...usedProperties(node)
-      }
-    }
-    tree.forEach(transformNode)
-
-    return [
-      Object.values(nodesMap).map(n => {
-        const relatedObservations = observations.filter(
-          o => o.emitterId === n.nodeId || o.receiverId === n.nodeId
-        )
-        const expectedMessages = {
-          Contributor: n.treeStructure.groupSize,
-          Leaf: n.treeStructure.fanout + 1,
-          Aggregator: n.treeStructure.fanout + 1,
-          Querier: n.treeStructure.groupSize + 1
-        }
-
-        return {
-          ...n,
-          id: n.nodeId,
-          startedWorking: relatedObservations.length > 0,
-          finishedWorking:
-            relatedObservations.length === expectedMessages[n.role]
-        }
-      }),
-      Object.entries(edgesMap)
-        .flatMap(([source, targets]) =>
-          targets.map(target => ({ source, target }))
-        )
-        .map(e => {
-          return {
-            ...e,
-            treeIndex:
-              nodesMap[e.target].treeIndex || nodesMap[e.source].treeIndex,
-            activeEdge: !!observations
-              ?.filter(o => o.action !== 'receiveShare')
-              ?.find(
-                o =>
-                  (o.receiverId === e.target && o.emitterId === e.source) ||
-                  (o.receiverId === e.source && o.emitterId === e.target)
-              )
-          }
-        })
-    ]
-  }, [observations, tree])
+  const setTreeNodes = useCallback(n => {
+    setTreeNodesState(n)
+  }, [])
+  const setTreeEdges = useCallback(n => {
+    setTreeEdgesState(n)
+  }, [])
 
   const handleLaunchExecution = useCallback(async () => {
     setIsRunning(true)
@@ -159,10 +74,6 @@ const Demonstration = () => {
     }
   }, [tree, treeStructure, supervisorWebhook, client.stackClient])
 
-  const handleRegenerateTree = useCallback(() => {
-    setTree(nodes && nodes.length > 0 ? createTree(treeStructure, nodes) : [])
-  }, [nodes, treeStructure])
-
   // Waiting for the execution to finish
   useEffect(() => {
     if (
@@ -184,46 +95,19 @@ const Demonstration = () => {
   return !nodes || isLoading ? (
     <Spinner size="xxlarge" middle />
   ) : (
-    <div>
-      <div className="full-agg-form">
-        <div>
-          <Label htmlFor="full-agg-contributors">Depth: </Label>
-          <Input
-            value={depth}
-            onChange={e => setDepth(Number(e.target.value))}
-            id="full-agg-contributors"
-          />
-        </div>
-        <div>
-          <Label htmlFor="full-agg-contributors">Fanout: </Label>
-          <Input
-            value={fanout}
-            onChange={e => setFanout(Number(e.target.value))}
-            id="full-agg-contributors"
-          />
-        </div>
-        <div>
-          <Label htmlFor="full-agg-contributors">Group Size: </Label>
-          <Input
-            value={groupSize}
-            onChange={e => setGroupSize(Number(e.target.value))}
-            id="full-agg-contributors"
-          />
-        </div>
-        <Button
-          variant="primary"
-          label="Regenerate tree"
-          onClick={handleRegenerateTree}
-          disabled={lastExecutionId !== tree[0]?.executionId || isRunning}
-        />
-        <Button
-          variant="primary"
-          label="Launch execution"
-          onClick={handleLaunchExecution}
-          busy={isRunning}
-          disabled={lastExecutionId === tree[0]?.executionId}
-        />
-      </div>
+    <div style={{ display: 'flex', flexDirection: 'column', gap: '1rem' }}>
+      <TreeCreator
+        busy={false}
+        setTreeNodes={setTreeNodes}
+        setTreeEdges={setTreeEdges}
+      />
+      <Button
+        variant="primary"
+        label="Launch execution"
+        onClick={handleLaunchExecution}
+        busy={isRunning}
+        disabled={lastExecutionId === tree[0]?.executionId}
+      />
       {treeNodes && treeEdges ? (
         <TreeNetworkGraph
           nodes={treeNodes}
@@ -232,7 +116,7 @@ const Demonstration = () => {
           height={400}
         />
       ) : null}
-      <NodesTable nodes={treeNodes} />
+      {treeNodes ? <NodesTable nodes={treeNodes} /> : null}
     </div>
   )
 }
